@@ -30,66 +32,45 @@
     "Issam Hakimi <issam.hakimi@googlemail.com>",
     "Marius Ursache <hello@mariusursache.com>",
     "Michael Lewkowitz <michael@igniter.com>",
-    "John Wehr <johnwehr@gmail.com>"
+    "John Wehr <johnwehr@gmail.com>",
+    "Ben James <benhjames@sky.com>"
   ],
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "https://raw.github.com/silentrob/superscript/master/LICENSE"
-    }
-  ],
+  "license": "MIT",
   "dependencies": {
-    "async": "^1.5.2",
-    "async-replace": "^1.0.1",
-    "bluebird": "^3.0.5",
-    "checksum": "^0.1.1",
+    "async": "^2.1.4",
     "commander": "^2.4.0",
-    "crc": "^3.0.0",
-    "debug": "^2.2.0",
+    "debug": "^2.6.1",
     "debug-levels": "^0.2.0",
-    "deepmerge": "^0.2.7",
-    "lemmer": "0.1.6",
-    "lodash": "^4.12.0",
-    "mkdirp": "^0.5.0",
-    "moment": "^2.13.0",
-    "mongodb": "^2.2.9",
-    "mongoose": "^4.5.10",
-    "mongoose-findorcreate": "^0.1.2",
-    "mongoose-path-tree": "^1.3.5",
+    "lodash": "^4.17.4",
+    "mkdirp": "^0.5.1",
+    "moment": "^2.17.0",
+    "mongo-tenant": "^1.0.3",
+    "mongoose": "4.7.8",
     "natural": "^0.4.0",
-    "node-normalizer": "^0.1.4",
-    "node-tense": "0.0.4",
-    "parts-of-speech": "^0.3.0",
-    "pluralize": "2.0.0",
-    "qtypes": "^0.1.6",
-    "re2": "^1.3.3",
-    "require-dir": "0.3.0",
-    "rhyme": "0.0.3",
-    "rimraf": "^2.4.4",
-    "roman-numerals": "~0.3.2",
-    "safe-eval": "^0.2.0",
-    "sfacts": "^0.2.0",
-    "ss-parser": "^0.5.0",
-    "string": "^3.3.1",
-    "syllablistic": "~0.1.0"
+    "pegjs": "^0.10.0",
+    "pluralize": "^3.0.0",
+    "require-dir": "^0.3.1",
+    "rhymes": "^1.0.1",
+    "roman-numerals": "^0.3.2",
+    "safe-eval": "^0.3.0",
+    "sfacts": "^1.0.1",
+    "ss-message": "^1.1.0",
+    "ss-parser": "^1.0.0",
+    "syllablistic": "^0.1.0",
+    "wordpos": "^1.1.2"
   },
   "devDependencies": {
-    "blanket": "~1.1.6",
-    "bluebird": "^2.9.34",
-    "coveralls": "^2.11.9",
-    "eslint": "^0.23.0",
-    "istanbul": "^0.4.3",
-    "mkdirp": "^0.5.0",
-    "mocha": "^3.0.2",
-    "mongodb": "^2.0.39",
-    "rimraf": "^2.4.2",
-    "rmdir": "^1.0.4",
-    "should": "^11.1.0"
-  },
-  "config": {
-    "blanket": {
-      "pattern": "/lib",
-      "data-cover-never": "node_modules"
-    }
+    "babel-cli": "^6.16.0",
+    "babel-preset-env": "^1.1.8",
+    "babel-register": "^6.18.0",
+    "coveralls": "^2.11.16",
+    "eslint": "^3.13.1",
+    "eslint-config-airbnb": "^14.0.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
+    "eslint-plugin-react": "^6.9.0",
+    "istanbul": "^1.1.0-alpha.1",
+    "mocha": "^3.2.0",
+    "should": "^11.2.0"
   }
 }
